project/edu/requirements.txt
django>=1.11
mongoengine==0.14.3

sphinx
nose
coverage

Pillow
django-bootstrap3

djangorestframework
django-rest-framework-mongoengine
blinker == 1.*
pymongo == 3.7.2
